Split up test_admin into files based on the endpoints they hit
